Split recording model classes to separate files
